+++ packages/client/ui/react-native/src/providers/CrossmintWalletProvider.tsx
@@ -29,9 +29,13 @@ export const CrossmintWalletEmailSignerContext = createContext<CrossmintWalletEm
 export interface CrossmintWalletProviderProps {
     children: ReactNode;
     createOnLogin?: CreateOnLogin;
+    callbacks?: {
+        onWalletCreationStart?: () => Promise<void>;
+        onTransactionStart?: () => Promise<void>;
+    };
 }
 
-export function CrossmintWalletProvider({ children, createOnLogin }: CrossmintWalletProviderProps) {
+export function CrossmintWalletProvider({ children, createOnLogin, callbacks }: CrossmintWalletProviderProps) {
     const { crossmint } = useCrossmint("CrossmintWalletProvider must be used within CrossmintProvider");
     const { apiKey, appId } = crossmint;
     const parsedAPIKey = validateAPIKey(apiKey);
@@ -160,6 +164,7 @@ export function CrossmintWalletProvider({ children, createOnLogin }: CrossmintWa
             createOnLogin={createOnLogin}
             onAuthRequired={onAuthRequired}
             _getEmailSignerIframe={_getEmailSignerIframe}
+            callbacks={callbacks}
         >
             <CrossmintWalletEmailSignerContext.Provider value={authContextValue}>
                 {children}